Moved urlsafekey class to entity
- updated test cases
